// microfont-poc/src/app/feature/Adminekyc/core/models/api-management.model.ts
export interface AdminApiManagement {
  id: number;
  name: string;
  key: string;
  port: string;
  url: string;
  user: string;
  password: string;
  credential: string;
 
}

export interface ApiManagementFormValue {
  name: string;
  key: string;
  port: string;
  url: string;
  user: string;
  password: string;
  credential: string;
}